feat(docs): add responsive tag list
